feat: create the crates/server and exclude quick-js from testing in the CI
