+++ .github/workflows/artifacts.yml
@@ -11,7 +11,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Test
-      run: cargo test --all
+      run: cargo test --workspace --exclude wasm-workers-quick-js-engine
   build:
     strategy:
       matrix:
