+++ .github/workflows/build.yml
@@ -43,7 +43,7 @@ jobs:
           target
         key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
     - name: Test
-      run: cargo test --all
+      run: cargo test --workspace --exclude wasm-workers-quick-js-engine
   build:
     runs-on: ubuntu-latest
     steps:
