+++ crates/server/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "wws-server"
+version = { workspace = true }
+edition = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+actix-web = { workspace = true }
+wws-data-kv = { workspace = true }
+wws-router = { workspace = true }
+wws-worker = { workspace = true }
+actix-files = "0.6.2"
