+++ crates/server/src/handlers/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2022 VMware, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod assets;
+pub mod not_found;
+pub mod worker;
